changes with service, table and repo

+++ CRUD-main/src/main/java/com/johnson/app/rest/Repository/UserRepository.java
@@ -2,10 +2,8 @@ package com.johnson.app.rest.Repository;
 
 import com.johnson.app.rest.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.math.BigInteger;
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, BigInteger> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByCommentTo(String commentTo);
 }
